requirements.txt
asgiref==3.8.1
django-tailwind==3.8.0
gunicorn==23.0.0
isort==5.13.2
packaging==24.2
sqlparse==0.5.3
typing_extensions==4.12.2
whitenoise==6.8.2